pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "movie-analysis-lakehouse"
version = "0.1.0"
description = "MovieLens 32M medallion lakehouse pipeline with Dagster, Spark, Glue, Athena, and QuickSight."
readme = "Readme.MD"
requires-python = ">=3.11"
dependencies = [
  "dagster==1.11.11",
  "dagster-webserver==1.11.11",
  "dagster-postgres==0.27.11",
  "dagster-aws==0.27.11",
  "pyspark==3.5.5",
  "delta-spark==3.2.0",
  "boto3==1.37.38",
  "python-dotenv==1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
]

[tool.setuptools.packages.find]
include = ["movie_pipeline*"]

[tool.dagster]
module_name = "movie_pipeline.definitions"
code_location_name = "movie_analysis"
